Tests/LibWeb/Text/expected/storage-estimate-does-not-break-session-storage.txt

window.open after estimate(): ok
sessionStorage after estimate(): value
